Added Event Location into CreateEvent activity
User presses "Add" to take them to a map screen. They enter a location using the AutoCompleteTextView, which puts  a marker down at that event location. When user presses "Save Location" this takes the text inside the textbox and saves it into string eventLocation. This string is passed to textview "locationText".